requirements.txt
Pandas
Numpy
langchain
openai
python-dotenv
langchain-openai
langchain_core